Moved source to 'Source' folder
Moved the project's source into its own folder
