--- AkDOS2/nbproject/private/private.properties
@@ -1,6 +1,0 @@
-compile.on.save=true
-do.depend=false
-do.jar=true
-javac.debug=true
-javadoc.preview=true
-user.properties.file=C:\\Users\\Armando\\AppData\\Roaming\\NetBeans\\8.1\\build.properties
